Split tests for scheme-utils.
