+++ client/lib/url/test/scheme-utils.js
@@ -1,0 +1,57 @@
+/**
+ * @jest-environment jsdom
+ */
+
+/**
+ * Internal dependencies
+ */
+import { addSchemeIfMissing, setUrlScheme } from '../scheme-utils';
+
+describe( 'addSchemeIfMissing()', () => {
+	test( 'should add scheme if missing', () => {
+		const source = 'example.com/path';
+		const expected = 'https://example.com/path';
+
+		const actual = addSchemeIfMissing( source, 'https' );
+
+		expect( actual ).toBe( expected );
+	} );
+
+	test( 'should skip if scheme exists', () => {
+		const source = 'https://example.com/path';
+		const expected = 'https://example.com/path';
+
+		const actual = addSchemeIfMissing( source, 'https' );
+
+		expect( actual ).toBe( expected );
+	} );
+} );
+
+describe( 'setUrlScheme()', () => {
+	test( 'should skip if scheme already set', () => {
+		const source = 'http://example.com/path';
+		const expected = 'http://example.com/path';
+
+		const actual = setUrlScheme( source, 'http' );
+
+		expect( actual ).toBe( expected );
+	} );
+
+	test( 'should add scheme if missing', () => {
+		const source = 'example.com/path';
+		const expected = 'http://example.com/path';
+
+		const actual = setUrlScheme( source, 'http' );
+
+		expect( actual ).toBe( expected );
+	} );
+
+	test( 'should replace scheme if different', () => {
+		const source = 'https://example.com/path';
+		const expected = 'http://example.com/path';
+
+		const actual = setUrlScheme( source, 'http' );
+
+		expect( actual ).toBe( expected );
+	} );
+} );
